M1/azor-chatdog/requirements.txt
google-genai
python-dotenv
llama-cpp-python
prompt_toolkit
colorama
pydantic>=2.0.0
fpdf2
markdown
anthropic